NAN(WiFi): Fix wrong variable size
NDPE Attribute is not dissected

Issue (and fix) reported by Darren Chen

Closed #17278

+++ epan/dissectors/packet-wifi-nan.c
@@ -1422,7 +1422,7 @@ dissect_attr_ndpe(proto_tree* attr_tree, tvbuff_t* tvb, gint offset, guint16 att
     offset += 4;
     dissected_len += 4;
     guint8 bits_type = tvb_get_bits8(tvb, offset * 8 + 4, 4);
-    guint8 bit_offset = (offset * 8) + 4;
+    guint32 bit_offset = (offset * 8) + 4;
     guint8 bits_status = tvb_get_bits8(tvb, bit_offset, 4);
 
     proto_tree_add_item(attr_tree, hf_nan_reason_code, tvb, offset + 1, 1, ENC_BIG_ENDIAN);
